refactor: Update module name
